+++ src/softpotato/analytical/kinetics/__init__.py
@@ -7,6 +7,5 @@ from .reversibility import nicholson_psi
 
 __all__ = [
     "catalytic_current",
-    "catalytic_current_ec_prime",
     "nicholson_psi",
 ]